unstructured_data/audio2text/requirements.txt
requests==2.32.3
snowflake-connector-python==3.14.0
snowflake-snowpark-python==1.29.1
click==8.1.8
whisper==1.1.10
ffmpeg-python==0.2.0
numpy==1.26.4
transformers==4.49.0
toml==0.10.2
evaluate==0.4.3
datasets==3.4.1
jiwer==3.1.0
whisper_normalizer==0.1.0
nemo_toolkit[asr]
